+++ app/page.tsx
@@ -3,15 +3,22 @@ import Button from "@/app/components/shared/button";
 import HeroSection from "@/app/components/shared/HeroSection";
 export default function Page() {
   return (
-    <main className="flex min-h-screen flex-col p-24">
+    <main>
       <HeroSection>
         <h1 className="text-white">Welcome to Alma</h1>
       </HeroSection>
-      <Link href="/dashboard">
-        <Button>
-          <h1>Lead Dashboard</h1>
-        </Button>
-      </Link>
+      <div className="mt-8 flex items-center justify-center">
+        <Link className="mx-auto" href="/dashboard/leads">
+          <Button>
+            <h1>Lead Dashboard</h1>
+          </Button>
+        </Link>
+        <Link className="mx-auto" href="/form">
+          <Button>
+            <h1>Get an Assessment</h1>
+          </Button>
+        </Link>
+      </div>
     </main>
   );
 }
